- Create a separate controller for categories and moving functionality of categories on it
- moving FileSettings class into Helper folder and making Games and GameStaions based on it

- Choosing Attribute folder for validations of cover and deleting Custom Validations folder

- moving GamesInStation action into GameStations Controller
