+++ GameZone/Areas/Admin/Controllers/GamesController.cs
@@ -21,35 +21,12 @@ namespace GameZone.Areas.Admin.Controllers
             _categoriesService = categoriesService;
             _devicesService = devicesService;
             _gamesService = gamesService;
+        }
 
         public IActionResult Index()
         {
             var games = _gamesService.GetAllGames();
             return View(games);
-        }
-
-        public async Task<IActionResult> GamesInStation(int stationId)
-        {
-            var station = await _context.GameStations
-                                           .Include(gs => gs.GameStationGames)
-                                           .ThenInclude(gsg => gsg.Game)
-                                           .ThenInclude(g => g.Category)
-                                           .FirstOrDefaultAsync(gs => gs.Id == stationId);
-
-            if (station is null)
-            {
-                return NotFound();
-            }
-
-            var gamesInStation = new GamesInStationVM
-            {
-                StationName = station.Name,
-                Games = station.GameStationGames
-                               .Select(gsg => gsg.Game)
-                               .ToList()
-            };
-
-            return View(gamesInStation);
         }
         
         [HttpGet]
@@ -145,85 +122,6 @@ namespace GameZone.Areas.Admin.Controllers
                 return Ok();
             else
                 return BadRequest();
-        }
-
-        public IActionResult AllGameCategories()
-        {
-            var categories = _context.Categories.ToList();
-            return View(categories);
-        }
-
-        [HttpGet]
-        public IActionResult CreateCategory()
-        {
-            return View();
-        }
-
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public IActionResult CreateCategory(GamesCategoriesVM model)
-        {
-            if (!ModelState.IsValid)
-                return View(model);
-
-            var category = new Category() { Name = model.Name };
-
-            _context.Categories.Add(category);
-            _context.SaveChanges();
-
-            return RedirectToAction(nameof(AllGameCategories));
-        }
-
-        [HttpGet]
-        public IActionResult EditCategory(int categoryId)
-        {
-            var category = _context.Categories.Find(categoryId);
-            if (category is null)
-                return NotFound();
-
-            var gamecategory = new GamesCategoriesVM { Id = categoryId, Name = category.Name };
-
-            return View(gamecategory);
-        }
-
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public IActionResult EditCategory(GamesCategoriesVM model)
-        {
-            if (!ModelState.IsValid)
-                return View(model);
-
-            var category = _context.Categories.Find(model.Id);
-            if (category is null)
-                return NotFound();
-
-            category.Name = model.Name;
-
-            _context.SaveChanges();
-
-            return RedirectToAction(nameof(AllGameCategories));
-        }
-
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public IActionResult DeleteCategory(int id)
-        {
-            var category = _context.Categories.Find(id);
-            if (category is null)
-                return NotFound();
-
-            _context.Categories.Remove(category);
-            _context.SaveChanges();
-
-            return Ok();
-        }
-
-        public IActionResult CheckIfNameExists(GamesCategoriesVM model)
-        {
-            var category = _context.Categories.SingleOrDefault(c => c.Name == model.Name);
-            var isAllowed = category is null || category.Id.Equals(model.Id);
-
-            return Json(isAllowed);
         }
     }
 }
