+++ GameZone/Areas/Admin/ViewModels/EditGameStationVM.cs
@@ -1,5 +1,5 @@
-﻿using GameZone.Constants;
-using GameZone.Custom_Validations;
+﻿using GameZone.Attributes;
+using GameZone.Helpers;
 using System.ComponentModel.DataAnnotations;
 
 namespace GameZone.Areas.Admin.ViewModels
@@ -13,9 +13,10 @@ namespace GameZone.Areas.Admin.ViewModels
 
 
         [Display(Name = "Image")]
+
         //validate extensions and cover
-        [AllowedExtensions(CoverSettings.allowedExtensions)]
-        [MaxFileSize(CoverSettings.maxFileSizeInBytes)]
+        [AllowedExtensions(FileSettings.AllowedExtensions)]
+        [MaxFileSize(FileSettings.MaxFileSizeInBytes)]
         public IFormFile? Cover { get; set; } = default!;
     }
 }
