+++ GameZone/Areas/Admin/Views/Categories/CreateCategory.cshtml
@@ -13,7 +13,7 @@
             <h2 class="mb-0">Creating New Category</h2>
         </div>
         <hr />
-        <form asp-controller="Games" asp-action="CreateCategory" method="post">
+        <form asp-controller="Categories" asp-action="CreateCategory" method="post">
 
             <div class="form-group mt-3">
                 <label for="loan_full_name" class="form-label" asp-for="Name"></label>
@@ -25,7 +25,7 @@
                 save
             </button>
 
-            <a href="/Admin/Games/AllGameCategories" class="text-white btn btn-outline-secondary" style="margin-left: 5px; margin-top: 1vw; width: 70px;">
+            <a href="/Admin/Categories/AllGameCategories" class="text-white btn btn-outline-secondary" style="margin-left: 5px; margin-top: 1vw; width: 70px;">
                 Cancel
             </a>
         </form>
