+++ GameZone/Areas/Admin/Views/Games/Create.cshtml
@@ -1,5 +1,4 @@
-﻿@using GameZone.Settings
-@model CreateGameFormVM
+﻿@model CreateGameFormVM
 
 @{
     ViewData["Title"] = "Create Game Page";
