+++ GameZone/Areas/Admin/Views/Games/Edit.cshtml
@@ -1,5 +1,4 @@
-﻿@using GameZone.Settings
-@model EditGameFormVM
+﻿@model EditGameFormVM
 
 @{
     ViewData["Title"] = "Edit Game Page";
