+++ GameZone/Areas/Admin/Views/Games/Index.cshtml
@@ -1,6 +1,4 @@
-﻿@using GameZone.Models
-@using GameZone.Settings
-@model IEnumerable<Game>
+﻿@model IEnumerable<Game>
 
 @{
     ViewData["Title"] = "All Games Page";
