+++ GameZone/Areas/Admin/Views/Shared/_Layout.cshtml
@@ -62,7 +62,7 @@
 
             <!-- Nav Item - Charts -->
             <li class="nav-item">
-                <a class="nav-link" asp-area="Admin" asp-controller="Games" asp-action="AllGameCategories">
+                <a class="nav-link" asp-area="Admin" asp-controller="Categories" asp-action="AllGameCategories">
                     <i class="fas fa-fw fa-chart-area"></i>
                     <span>Games Categories</span>
                 </a>
