+++ GameZone/Areas/Admin/Views/_ViewImports.cshtml
@@ -1,6 +1,5 @@
 ﻿@using GameZone
 @using GameZone.Areas.Admin.ViewModels
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @using GameZone.Models;
-@using GameZone.Constants
-@using GameZone.Areas.Admin.ViewModels;
+@using GameZone.Helpers
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
