--- GameZone/Constants/CoverSettings.cs
@@ -1,11 +1,0 @@
-﻿namespace GameZone.Constants
-{
-    public class CoverSettings
-    {
-        public const string stationFilePath = "/Assets/Images/StationsImages";
-        public const string gamesImagePath = "/Assets/Images/GamesImages+2";
-        public const string allowedExtensions = ".jpg,.jpeg,.png";
-        public const int maxFileSizeInMB = 1;
-        public const int maxFileSizeInBytes = maxFileSizeInMB * 1024 * 1024;
-    }
-}
