+++ GameZone/Views/Games/GetAll.cshtml
@@ -53,7 +53,7 @@
                         <div class="card bg-transparent border-0 text-center">
 
                             <div class="card-img">
-                                <img src="@CoverSettings.gamesImagePath/@game.Cover" alt="Cover Image" class="rounded w-100 gray-image" width="300" height="332">
+                                <img src="@FileSettings.GamesImagesPath/@game.Cover" alt="Cover Image" class="rounded w-100 gray-image" width="300" height="332">
                                 <a class="hidden-button bg-primary rounded-2" asp-controller="Games" asp-action="GameDetails" asp-route-gameId="@game.Id">
                                     Learn More
                                 </a>
