+++ GameZone/Views/_ViewImports.cshtml
@@ -1,5 +1,5 @@
 ﻿@using GameZone
 @using GameZone.Models
 @using GameZone.ViewModels.Games
-@using GameZone.Constants
+@using GameZone.Helpers
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
